server/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29

# Core ML/Data deps
numpy>=1.24
pandas>=2.2
scikit-learn>=1.4

# HTTP clients
requests>=2.31
httpx>=0.27

# Validation + forms
python-dotenv>=1.0.0
email-validator>=2.1

# Security (AUTH)
passlib[argon2]>=1.7
python-jose[cryptography]>=3.3

# File handling
aiofiles>=23.2.1

# Testing
pytest>=8.1

# Vector search removed (FAISS not supported on Python 3.14)
# databases/sqlalchemy (if still used in your project)
sqlalchemy>=2.0.32
databases>=0.7.3
